Modularize: Export getHolidaysOnDate() without HebrewCalendar
