+++ size-demo/sizes.md
@@ -1,7 +1,7 @@
 # Bundle sizes
 
-- `getHoliday.mjs`: 409 kB
-- `getHoliday.min.mjs`: 147 kB
+- `getHoliday.mjs`: 132 kB
+- `getHoliday.min.mjs`: 45 kB
 - `parshiyot.mjs`: 409 kB
 - `parshiyot.min.mjs`: 147 kB
 - `sedra.mjs`: 89.2 kB
